requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
reportlab==4.0.7
python-multipart==0.0.6
httpx==0.24.1